Update game.py
In this updated code, after each day, the user is required to press Enter to continue to the next day. The game runs for three days, and after the third day, the user is asked if they would like to continue playing. If the user chooses “yes,” the main() function is called again to start a new game. If the user chooses “no,” the game ends with a message.